Extract BenchView helpers to utilities so tests that don't use CPC can call them
